0.2.5
Applied multithreading for uploading large files to prevent unresponsiveness.
Implemented saving of port on the server.
added a link to the github page :)
Attempting to fix the client progress bar

+++ OpenFTP Server/headers/ftp_manager.h
@@ -55,7 +55,7 @@ public:
 	static void deleteFiles(const QJsonArray& filesToDelete);
 	static bool renameFile(const QString& filePath, const QString& oldFileName, QString& newFileName);
 	static bool createFolder(const QString& newFolderName);
-	static Transfer startFileUpload(const int& userIndex,const QString& fileName, const QString& filePath, const int& fileSize, const bool& baseDir, const QString& directoryToReturn);
+	static Transfer startFileUpload(const int& userIndex,const QString& fileName, const QString& filePath, const qint64& fileSize, const bool& baseDir, const QString& directoryToReturn);
 	static Transfer createPendingFileDownload(const int& userIndex, const QString& filePath, const QString& fileName, const bool& baseDir, QString& errorString);
 	static bool beginFileDownload(const Transfer& download, QTcpSocket* socket, QString& errorString);
 	static int processFileUpload(const QByteArray& data, Transfer& upload);
