+++ OpenFTP Server/headers/transfer.h
@@ -6,7 +6,7 @@ class Transfer
 {
 
 public:
-	Transfer(int _userIndex,QString _fileName, QString _filePath, int _fileSize, bool _isBaseDir, QString _directoryToReturn);
+	Transfer(int _userIndex,QString _fileName, QString _filePath, qint64 _fileSize, bool _isBaseDir, QString _directoryToReturn);
 	Transfer(int _userIndex, QString _fileName, QString _filePath, bool _isBaseDir);
 	Transfer();
 
@@ -15,11 +15,11 @@ public:
 	void cancelUpload();
 
 	int userIndex = -1;
-	int writtenBytes = 0;
+	qint64 writtenBytes = 0;
 	QString fileName;
 	QString filePath;
 	bool isBaseDir;
-	int fileSize = -1;
+	qint64 fileSize = -1;
 	QString directoryToReturn;
 	int numOfPacketsSent = 0;
